chore: Split Windows impl to a separate module
