+++ src/lib.rs
@@ -28,24 +28,24 @@
 #![deny(rust_2018_idioms)]
 #![allow(deprecated)] // uses env::home_dir when necessary
 
+mod windows;
+
 use std::env;
 use std::io;
 use std::path::{Path, PathBuf};
-#[cfg(windows)]
-use winapi::shared::minwindef::DWORD;
 
 /// Returns the path of the current user's home directory if known.
 ///
 /// # Unix
 ///
-/// Returns the value of the 'HOME' environment variable if it is set
+/// Returns the value of the `HOME` environment variable if it is set
 /// and not equal to the empty string. Otherwise, it tries to determine the
 /// home directory by invoking the `getpwuid_r` function on the UID of the
 /// current user.
 ///
 /// # Windows
 ///
-/// Returns the value of the 'USERPROFILE' environment variable if it
+/// Returns the value of the `USERPROFILE` environment variable if it
 /// is set and not equal to the empty string. If both do not exist,
 /// [`GetUserProfileDirectory`][msdn] is used to return the
 /// appropriate path.
@@ -55,112 +55,20 @@ use winapi::shared::minwindef::DWORD;
 /// # Examples
 ///
 /// ```
-/// use std::env;
-///
-/// match env::home_dir() {
+/// match home::home_dir() {
 ///     Some(path) => println!("{}", path.display()),
 ///     None => println!("Impossible to get your home dir!"),
 /// }
 /// ```
 pub fn home_dir() -> Option<PathBuf> {
-    home_dir_()
+    home_dir_inner()
 }
 
 #[cfg(windows)]
-fn home_dir_() -> Option<PathBuf> {
-    use std::ptr;
-    use winapi::shared::winerror::ERROR_INSUFFICIENT_BUFFER;
-    use winapi::um::errhandlingapi::GetLastError;
-    use winapi::um::handleapi::CloseHandle;
-    use winapi::um::processthreadsapi::{GetCurrentProcess, OpenProcessToken};
-    use winapi::um::userenv::GetUserProfileDirectoryW;
-    use winapi::um::winnt::TOKEN_READ;
-
-    env::var_os("USERPROFILE")
-        .map(PathBuf::from)
-        .or_else(|| unsafe {
-            let me = GetCurrentProcess();
-            let mut token = ptr::null_mut();
-            if OpenProcessToken(me, TOKEN_READ, &mut token) == 0 {
-                return None;
-            }
-            let _g = scopeguard::guard(token, |h| {
-                let _ = CloseHandle(h);
-            });
-            fill_utf16_buf(
-                |buf, mut sz| {
-                    match GetUserProfileDirectoryW(token, buf, &mut sz) {
-                        0 if GetLastError() != ERROR_INSUFFICIENT_BUFFER => 0,
-                        0 => sz,
-                        _ => sz - 1, // sz includes the null terminator
-                    }
-                },
-                os2path,
-            )
-            .ok()
-        })
-}
-
-#[cfg(windows)]
-fn os2path(s: &[u16]) -> PathBuf {
-    use std::ffi::OsString;
-    use std::os::windows::ffi::OsStringExt;
-    PathBuf::from(OsString::from_wide(s))
-}
-
-#[cfg(windows)]
-fn fill_utf16_buf<F1, F2, T>(mut f1: F1, f2: F2) -> io::Result<T>
-where
-    F1: FnMut(*mut u16, DWORD) -> DWORD,
-    F2: FnOnce(&[u16]) -> T,
-{
-    use winapi::shared::winerror::ERROR_INSUFFICIENT_BUFFER;
-    use winapi::um::errhandlingapi::{GetLastError, SetLastError};
-
-    // Start off with a stack buf but then spill over to the heap if we end up
-    // needing more space.
-    let mut stack_buf = [0u16; 512];
-    let mut heap_buf = Vec::new();
-    unsafe {
-        let mut n = stack_buf.len();
-        loop {
-            let buf = if n <= stack_buf.len() {
-                &mut stack_buf[..]
-            } else {
-                let extra = n - heap_buf.len();
-                heap_buf.reserve(extra);
-                heap_buf.set_len(n);
-                &mut heap_buf[..]
-            };
-
-            // This function is typically called on windows API functions which
-            // will return the correct length of the string, but these functions
-            // also return the `0` on error. In some cases, however, the
-            // returned "correct length" may actually be 0!
-            //
-            // To handle this case we call `SetLastError` to reset it to 0 and
-            // then check it again if we get the "0 error value". If the "last
-            // error" is still 0 then we interpret it as a 0 length buffer and
-            // not an actual error.
-            SetLastError(0);
-            let k = match f1(buf.as_mut_ptr(), n as DWORD) {
-                0 if GetLastError() == 0 => 0,
-                0 => return Err(io::Error::last_os_error()),
-                n => n,
-            } as usize;
-            if k == n && GetLastError() == ERROR_INSUFFICIENT_BUFFER {
-                n *= 2;
-            } else if k >= n {
-                n = k;
-            } else {
-                return Ok(f2(&buf[..k]));
-            }
-        }
-    }
-}
+use windows::home_dir_inner;
 
 #[cfg(any(unix, target_os = "redox"))]
-fn home_dir_() -> Option<PathBuf> {
+fn home_dir_inner() -> Option<PathBuf> {
     env::home_dir()
 }
